Refactor BMG files reading
